fix: skip RSVP write on no-op resubmit to preserve updated_at
update_or_create was writing on every submission, bumping updated_at
even when status/plus_one were unchanged. Now short-circuits when the
resolved RSVP matches the existing row.

Also splits TestCreateEventWithCohosts out of test_rsvp.py into
test_event_cohosts.py to stay under the 500-line file limit.

(Issue 948)
